Add files via upload
Data Preprocessing:
In Data Preprocessing we check whether there are any duplicate values, missing values and we deal with them initally. In the Dataset I have found and removed them. As we see from the attribute information, Value of HelpfulnessDenominator should be always greater than value of HelpfulnessNumerator, because HelpfulnessDenominator is the sum of both useful review and not useful review while HelpfulnessNumerator is only the number of users who found the review helpful. So removing records or datapoints that doesnt satisfy this condition.
we are also removing the neutral reviews(3 star reviews) from the dataset for better predicitons, considering 1 & 2 star reviews as negative reviews and 4 % 5 star as positive reviews.
Text Preprocessing: Stemming, Stop-words removal, HTML tags removal, Punctuation marks removal
Now that we have finished deduplication our data requires some preprocessing before we go on further with analysis and making the prediction model.
Hence in the Preprocessing phase we do the following in the order below:-
1.Begin by removing the html tags
2.Remove any punctuations or limited set of special characters like , or . or # etc.
3.Check if the word is made up of english letters and is not alpha-numeric
4.Check to see if the length of the word is greater than 2 (as it was researched that there is no adjective in 2-letters)
5.Convert the word to lowercase
6.Remove Stopwords
7.Finally Snowball Stemming the word (it was obsereved to be better than Porter Stemming)
8.After which we collect the words used to describe positive and negative reviews
9.Converting Text into Numerical vectors.
    1.Bag of Words
    2.Term Frequency Inverse Document Frequency (TF-IDF)
    3.Average Word2Vec
    4.TF_IDF Word2Vec (You can also find the optimal codes)